feature: update employee

+++ src/main/java/com/company_management/dto/response/project/ResponseProjectDashboardTO.java
@@ -13,5 +13,5 @@ public class ResponseProjectDashboardTO {
     private String projectName;
     private int numberOfTasks;
     private int numberOfTasksDone;
-    private long percentage;
+    private Double percentage;
 }
